… LNZWeakCollection/Extensions/Algorithms.swift