# Issue
The implementation of `cleanup` runs on O(n^3) times.

# Why?

The `cleanup` function is implemented with 2 operations: `index(where:)` and `remove(at:)`. Both of these functions are O(n). There is also a O(n) loop.
The implementation is:  
```swift
while let index = weakReferecesContainers.index(where: { $0.weakReference == nil }) {
    weakReferecesContainers.remove(at: index)
    if index < currentIndex {
        currentIndex -= 1
    }
}
```

An O(n) operation is iteratively computed in an O(n) loop. Another O(n) operation is done inside this O(n^2) loop.

# Solution

I recommend changing the code to:

```swift
let count = weakReferecesContainers.count
let last = weakReferecesContainers.stablePartition(isSuffixElement: { $0 == nil })
let nilCount = count - last + 1
weakReferecesContainers.removeLast(nilCount)
```

> A side note, `weakReferecesContainers` should be renamed to `weakReferenceContainers`. I believe that the typo was not intentional.

# Benchmarking

**Old method**

| Elements | Time (s) |
| --------- | -------- |
| 100          | 0.0003  |
| 1,000       | 0.0196   |
| 10,000     | 1.46       |
| 50,000    | 35.5       |

**New method**

| Elements | Time (s) |
| --------- | -------- |
| 100          | 0.0008  |
| 1,000       | 0.0056  |
| 10,000     | 0.0606  |
| 50,000    | 0.292     |

**Change**

| Elements | Old        | New       | Improvement |
| --------- | -------- | -------- | ------------- |
| 100          | 0.0003  | 0.0008   | -62.5%           |
| 1,000       | 0.0196  | 0.0056   | 250%              |
| 10,000     | 1.46      | 0.0606   | 2309%            |
| 50,000     | 35.5     | 0.292      | 12067%          |

> The tests were run using XCTests.